+++ src/actions2.lisp
@@ -3,7 +3,9 @@
   (:export
    #:command
    #:render-command
-   #:add-command))
+   #:add-command
+   #:on-missing-action
+   #:eval-action))
 (in-package weblocks.actions)
 
 
@@ -44,3 +46,32 @@ After action processing these commands will be sent for execution on the client.
     reset-commands-list ()
   (let (*commands*)
     (weblocks.hooks:call-next-hook)))
+
+
+(defgeneric on-missing-action (app action-name)
+  (:documentation "Must be overridden by application to prevent default
+behaviour - redirect to a root of the application.
+The new method should determine the behavior in this
+situation (e.g. redirect, signal an error, etc.)."))
+
+
+(defmethod on-missing-action (app action-name)
+  (declare (ignorable app action-name))
+  (weblocks::redirect
+   (weblocks::make-webapp-uri "/" app)))
+
+
+(defgeneric eval-action (app action-name arguments)
+  (:documentation "Evaluates the action that came with the request."))
+
+
+(defmethod eval-action (app action-name arguments)
+  "Evaluates the action that came with the request."
+  (let ((action (weblocks::get-request-action action-name)))
+
+    (unless action
+      (on-missing-action app action-name))
+    
+    (log:debug "Calling" action "with" arguments "and" action-name)
+    (weblocks::safe-apply action
+                          arguments)))